update decimal prices

+++ api/src/db/scripts/create_menu_table.sql
@@ -2,5 +2,5 @@ CREATE TABLE IF NOT EXISTS restaurant.menu (
     id INT AUTO_INCREMENT PRIMARY KEY,
     type VARCHAR(100) NOT NULL,
     name VARCHAR(100) NOT NULL,
-    price SMALLINT unsigned NOT NULL
+    price DECIMAL( 4, 2 ) NOT NULL
 ) ENGINE=InnoDB;
